backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
httpx==0.26.0
python-jose==3.3.0
python-multipart==0.0.6
